# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/trips": {
      "post": {
        "operationId": "TripController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Viagem criada com sucesso"
          },
          "403": {
            "description": "Acesso negado — somente motoristas"
          },
          "409": {
            "description": "Motorista já possui viagem ativa"
          }
        },
        "summary": "Iniciar nova viagem (OUTBOUND ou RETURN)",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{id}/end": {
      "patch": {
        "operationId": "TripController_end",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem encerrada com sucesso"
          },
          "400": {
            "description": "Viagem não está ativa"
          },
          "404": {
            "description": "Viagem não encontrada"
          }
        },
        "summary": "Encerrar viagem ativa",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/api/v1/trips/{id}/students": {
      "get": {
        "description": "Contrato declarado — implementação na Story 3.5a.",
        "operationId": "TripController_getStudents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da viagem",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de alunos da viagem",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TripStudentsResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — somente motoristas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "TRIP_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "NOT_IMPLEMENTED — contrato declarado, implementação na Story 3.5a",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar alunos da viagem com status de embarque",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/api/v1/boarding/check-in": {
      "post": {
        "description": "Contrato declarado — implementação na Story 3.3a. Idempotência: mesma X-Idempotency-Key retorna o resultado anterior com 201, sem duplicar.",
        "operationId": "BoardingController_checkIn",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "UUID v4 gerado no cliente — chave da fila offline (Architecture §5, Tier 2). Mesma key reenviada retorna o mesmo resultado sem duplicar o check-in.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Check-in registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CheckInResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "INVALID_QR_CODE — studentId ou tripId ausente/malformado. O QR bruto nunca trafega (o app do motorista decodifica e envia campos estruturados), então este erro é validação de shape do body, não do QR em si. MISSING_IDEMPOTENCY_KEY — header X-Idempotency-Key ausente ou vazio (é required).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "STUDENT_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "TRIP_NOT_ACTIVE | DUPLICATE_CHECK_IN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "NOT_IMPLEMENTED — contrato declarado, implementação na Story 3.3a",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar check-in de embarque de um aluno",
        "tags": [
          "boarding"
        ]
      }
    }
  },
  "info": {
    "title": "PureUrban",
    "description": "PureUrban Official API Documentation",
    "version": "0.2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "description": "ID da rota",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "type": {
            "type": "string",
            "enum": [
              "OUTBOUND",
              "RETURN"
            ],
            "description": "Tipo de viagem: ida (OUTBOUND) ou volta (RETURN)"
          },
          "relatedTripId": {
            "type": "string",
            "description": "ID da viagem de ida (obrigatório para RETURN)",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          }
        },
        "required": [
          "routeId",
          "type"
        ]
      },
      "TripStudentItemDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "description": "ID do aluno",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "Nome do aluno",
            "example": "Maria Silva"
          },
          "status": {
            "type": "string",
            "enum": [
              "CHECKED_IN",
              "NOT_CHECKED_IN",
              "NOT_RETURNING"
            ],
            "description": "Status de embarque do aluno na viagem",
            "example": "NOT_CHECKED_IN"
          },
          "checkedInAt": {
            "type": "string",
            "description": "Data e hora do check-in (ISO 8601 UTC), null se ainda não embarcou",
            "example": "2026-07-12T21:10:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "studentId",
          "name",
          "status",
          "checkedInAt"
        ]
      },
      "BoardingSummaryDto": {
        "type": "object",
        "properties": {
          "boarded": {
            "type": "number",
            "description": "Quantidade de alunos já embarcados",
            "example": 12
          },
          "total": {
            "type": "number",
            "description": "Quantidade total de alunos na viagem",
            "example": 20
          }
        },
        "required": [
          "boarded",
          "total"
        ]
      },
      "TripStudentsResponseDto": {
        "type": "object",
        "properties": {
          "students": {
            "description": "Lista de alunos da viagem com status de embarque",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripStudentItemDto"
            }
          },
          "summary": {
            "description": "Resumo agregado de embarque — sempre calculado no servidor",
            "allOf": [
              {
                "$ref": "#/components/schemas/BoardingSummaryDto"
              }
            ]
          }
        },
        "required": [
          "students",
          "summary"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Código de erro tipado",
            "example": "TRIP_NOT_ACTIVE"
          },
          "message": {
            "type": "string",
            "description": "Mensagem legível descrevendo o erro",
            "example": "A viagem não está ativa"
          },
          "details": {
            "type": "object",
            "description": "Detalhes adicionais do erro (opcional)"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CheckInResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do registro de check-in",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "studentId": {
            "type": "string",
            "description": "ID do aluno que fez check-in",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "tripId": {
            "type": "string",
            "description": "ID da viagem em que o check-in ocorreu",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "checkedInAt": {
            "type": "string",
            "description": "Data e hora do check-in (ISO 8601 UTC)",
            "example": "2026-07-12T21:10:00.000Z"
          },
          "status": {
            "type": "string",
            "description": "Status do check-in",
            "enum": [
              "CHECKED_IN"
            ],
            "example": "CHECKED_IN"
          }
        },
        "required": [
          "id",
          "studentId",
          "tripId",
          "checkedInAt",
          "status"
        ]
      },
      "CheckInRequestDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "description": "ID do aluno decodificado do QR code",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "tripId": {
            "type": "string",
            "description": "ID da viagem ativa em que o check-in ocorre",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          }
        },
        "required": [
          "studentId",
          "tripId"
        ]
      }
    }
  }
}